Adds index-cards and uses them for both workspaces and result-forms

+++ src/main/webapp/confirm-delete/confirm-delete.js
@@ -41,7 +41,9 @@ const ConfirmDelete = props => {
               color="secondary"
               variant="contained"
               onClick={() => {
-                onDelete()
+                if (typeof onDelete === 'function') {
+                  onDelete()
+                }
                 setConfirmDelete(false)
               }}
             >
@@ -53,6 +55,7 @@ const ConfirmDelete = props => {
       <IconButton
         onClick={e => {
           e.stopPropagation()
+          e.preventDefault()
           setConfirmDelete(true)
         }}
       >
